removed references to Vbe.Interop in Rubberduck.Parsing
