+++ Rubberduck.VBEEditor/SafeComWrappers/VBA/CodeModule.cs
@@ -90,79 +90,6 @@ namespace Rubberduck.VBEditor.SafeComWrappers.VBA
             Target.DeleteLines(1, CountOfLines);
         }
 
-        public static string[] GetSanitizedCode(ICodeModule module)
-        {
-            var lines = module.CountOfLines;
-            if (lines == 0)
-            {
-                return new string[] { };
-            }
-
-            var code = module.GetLines(1, lines).Replace("\r", string.Empty).Split('\n');
-
-            StripLineNumbers(code);
-            return code;
-        }
-
-        /// <summary>
-        /// Gets an array of strings where each element is a line of code in the Module,
-        /// with line numbers stripped and any other pre-processing that needs to be done.
-        /// </summary>
-        public string[] GetSanitizedCode()
-        {
-            var lines = CountOfLines;
-            if (lines == 0)
-            {
-                return new string[] { };
-            }
-
-            var code = GetLines(1, lines).Replace("\r", string.Empty).Split('\n');
-
-            StripLineNumbers(code);
-            return code;
-        }
-
-        private static void StripLineNumbers(string[] lines)
-        {
-            var continuing = false;
-            for (var line = 0; line < lines.Length; line++)
-            {
-                var code = lines[line];
-                int? lineNumber;
-                if (!continuing && HasNumberedLine(code, out lineNumber))
-                {
-                    var lineNumberLength = lineNumber.ToString().Length;
-                    if (lines[line].Length > lineNumberLength)
-                    {
-                        // replace line number with as many spaces as characters taken, to avoid shifting the tokens
-                        lines[line] = new string(' ', lineNumberLength) + code.Substring(lineNumber.ToString().Length + 1);
-                    }
-                }
-
-                continuing = code.EndsWith("_");
-            }
-        }
-
-        private static bool HasNumberedLine(string codeLine, out int? lineNumber)
-        {
-            lineNumber = null;
-
-            if (string.IsNullOrWhiteSpace(codeLine.Trim()))
-            {
-                return false;
-            }
-
-            int line;
-            var firstToken = codeLine.TrimStart().Split(' ')[0];
-            if (int.TryParse(firstToken, out line))
-            {
-                lineNumber = line;
-                return true;
-            }
-
-            return false;
-        }
-
         private string _previousContentHash;
         public string ContentHash()
         {
